Bevy eighteen (#132)
* placeholder branch init...

* bump deps, just waiting on bevy_egui to update themselves...

* chore: bump version, toml versions to support bevy 0.18.x

* chore: changes to irradiance api? -- just make it compile again.

+++ src/system/config.rs
@@ -154,8 +154,8 @@ impl UserSession {
     fn pop_gif_buffer(&mut self, images: &mut ResMut<Assets<Image>>) -> anyhow::Result<Image> {
         let (width, height) = self.window_dims;
         let size = Extent3d {
-            width: width as u32,
-            height: height as u32,
+            width: width,
+            height: height,
             ..Default::default()
         };
